refactor: split magic constants passes
